Rename and move skinnable line component to a more commomn place
